requirements.txt
pandas
numpy
xgboost
supabase
python-dotenv
MLB-StatsAPI
requests
scikit-learn
